[django 1.7] Added AppConfig loaders #216
* notifications
* metrics
* sync
* ui
