+++ nodeshot/community/notifications/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'nodeshot.community.notifications.apps.AppConfig'
