+++ nodeshot/interop/sync/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'nodeshot.interop.sync.apps.AppConfig'
