+++ nodeshot/ui/default/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'nodeshot.ui.default.apps.AppConfig'
